+++ demos/automerge-repo-todos/src/hooks/useAuth.tsx
@@ -1,9 +1,9 @@
 import { useContext } from 'react'
-import { authContext } from '../components/LocalFirstAuthProvider'
+import { LocalFirstAuthContext } from '../components/LocalFirstAuthProvider'
 
 // Convenience wrapper around our authContext for accessing the auth data and provider
 export const useAuth = () => {
-  const context = useContext(authContext)
+  const context = useContext(LocalFirstAuthContext)
   if (context === null) throw new Error('useAuth must be used within a LocalFirstAuthProvider')
   return context
 }
